requirements.txt
amqp==2.5.2
asgiref==3.2.7
billiard==3.6.3.0
cachetools==4.1.0
celery==4.4.2
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
Django==3.0.7
django-cors-headers==3.3.0
djangorestframework==3.11.0
djangorestframework-simplejwt==4.4.0
docker==4.2.1
future==0.18.2
google-api-core==1.20.0
google-api-python-client==1.9.3
google-auth==1.17.2
google-auth-httplib2==0.0.3
googleapis-common-protos==1.52.0
gunicorn==20.0.4
httplib2==0.18.1
idna==2.9
importlib-metadata==1.6.0
Jinja2==2.11.2
joblib==0.14.1
kombu==4.6.8
livereload==2.6.1
lunr==0.5.6
Markdown==3.2.1
MarkupSafe==1.1.1
mkautodoc==0.1.0
mkdocs==1.1
nltk==3.5
numpy==1.18.4
pandas==1.0.3
protobuf==3.12.2
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyJWT==1.7.1
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
redis==3.5.2
regex==2020.4.4
requests==2.23.0
rsa==4.6
six==1.14.0
sqlparse==0.3.1
tornado==6.0.4
tqdm==4.46.0
uritemplate==3.0.1
urllib3==1.25.9
vine==1.3.0
websocket-client==0.57.0
xlrd==1.2.0
zipp==3.1.0
